se organiza las restricciones en los servicios, se agregran funcionalidades y restricciones de usuario del blog

+++ proyectoweb/servicios/views.py
@@ -1,3 +1,4 @@
+from django.contrib import messages
 from django.shortcuts import redirect, render, get_object_or_404
 from servicios.models import Servicio
 from .forms import ServicioForm
@@ -11,33 +12,36 @@ def servicios(request):
 
 @login_required
 def agregar_servicio(request):
-    if request.method == 'POST':
-        form = ServicioForm(request.POST, request.FILES)
-        if form.is_valid():
-            form.save()
-            return redirect('servicios')
-    else:
-        form = ServicioForm()
+    if request.user.is_superuser:
+        if request.method == 'POST':
+            form = ServicioForm(request.POST, request.FILES)
+            if form.is_valid():
+                form.save()
+                return redirect('servicios')
+        else:
+            form = ServicioForm()
 
-    context = {'form':form}
-    return render(request, 'servicios/agregar_servicio.html', context)
+        context = {'form':form}
+        return render(request, 'servicios/agregar_servicio.html', context)
 
 @login_required
 def editar_servicio(request, id):
-    servicio = get_object_or_404(Servicio, pk=id)
-    if request.method == 'POST':
-        form = ServicioForm(request.POST, instance=servicio)
-        if form.is_valid():
-            form.save()
-            return redirect('servicios')
-    else:
-        form = ServicioForm(instance=servicio)
-    
-    context = {'form':form}
-    return render(request,'servicios/editar_servicio.html', context)
+    if request.user.is_superuser:
+        servicio = get_object_or_404(Servicio, pk=id)
+        if request.method == 'POST':
+            form = ServicioForm(request.POST, instance=servicio)
+            if form.is_valid():
+                form.save()
+                return redirect('servicios')
+        else:
+            form = ServicioForm(instance=servicio)
+        
+        context = {'form':form}
+        return render(request,'servicios/editar_servicio.html', context)
 
 @login_required
 def eliminar_servicio(request, id):
-    servicio = get_object_or_404(Servicio, pk=id)
-    servicio.delete()
-    return redirect('servicios')
+    if request.user.is_superuser:
+        servicio = get_object_or_404(Servicio, pk=id)
+        servicio.delete()
+        return redirect('servicios')
